+++ apps/pushnotif-func/src/adapters/cosmos/massive-jobs.ts
@@ -1,14 +1,10 @@
-import { Container, Database, ErrorResponse } from "@azure/cosmos";
+import { Container, ErrorResponse } from "@azure/cosmos";
 
 import { ErrorInternal, ErrorNotFound } from "../../domain/error";
 import { MassiveJob, MassiveJobsRepository } from "../../domain/massive-jobs";
 
 export class CosmosMassiveJobsAdapter implements MassiveJobsRepository {
-  container: Container;
-
-  constructor(cosmosdbInstance: Database, containerName: string) {
-    this.container = cosmosdbInstance.container(containerName);
-  }
+  constructor(private container: Container) {}
 
   async createMassiveJob(job: MassiveJob) {
     try {
@@ -16,24 +12,6 @@ export class CosmosMassiveJobsAdapter implements MassiveJobsRepository {
       return result.item.id;
     } catch (err) {
       return new ErrorInternal("Failed to create massive job", err);
-    }
-  }
-
-  async updateMassiveJob(job: MassiveJob) {
-    try {
-      const result = await this.container.item(job.id).replace(job);
-      return result.item.id;
-    } catch (err) {
-      if (err instanceof ErrorResponse) {
-        switch (err.code) {
-          case 404:
-            return new ErrorNotFound("Massive job not found", err);
-          default:
-            return new ErrorInternal("Failed to update massive job", err);
-        }
-      }
-
-      return new ErrorInternal("Failed to update massive job", err);
     }
   }
 
@@ -57,6 +35,24 @@ export class CosmosMassiveJobsAdapter implements MassiveJobsRepository {
       }
 
       return new ErrorInternal("Failed to get massive job", err);
+    }
+  }
+
+  async updateMassiveJob(job: MassiveJob) {
+    try {
+      const result = await this.container.item(job.id).replace(job);
+      return result.item.id;
+    } catch (err) {
+      if (err instanceof ErrorResponse) {
+        switch (err.code) {
+          case 404:
+            return new ErrorNotFound("Massive job not found", err);
+          default:
+            return new ErrorInternal("Failed to update massive job", err);
+        }
+      }
+
+      return new ErrorInternal("Failed to update massive job", err);
     }
   }
 }