+++ apps/pushnotif-func/src/domain/health.ts
@@ -1,0 +1,3 @@
+import { ErrorInternal } from "./error";
+
+export type HealthCheck = () => Promise<ErrorInternal | undefined>;