+++ apps/pushnotif-func/src/main.ts
@@ -29,19 +29,25 @@ import { pipe } from "fp-ts/lib/function";
 import nodeFetch from "node-fetch";
 
 import { TelemetryClient } from "./adapters/appinsights";
+import { blobServiceHealthcheck } from "./adapters/blob-service/health";
 import {
   Config,
   configFromEnvironment,
   loadConfigFromEnvironment,
 } from "./adapters/config";
+import { cosmosHealthcheck } from "./adapters/cosmos/health";
 import { CosmosInstallationSummaryAdapter } from "./adapters/cosmos/installation";
 import { CosmosMassiveJobsAdapter } from "./adapters/cosmos/massive-jobs";
 import { createMassiveNotificationJobHandler } from "./adapters/functions/create-massive-notification-job";
+import { getHealthHandler } from "./adapters/functions/health";
 import { getInfoHandler } from "./adapters/functions/info";
 import getUpdateInstallationHandler from "./adapters/functions/update-installation";
 import getInstallationUpdateDispatcher from "./adapters/functions/update-installation-dispatch";
+import { notificationHubHealthcheck } from "./adapters/notification-hub/health";
 import { NotificationHubInstallationAdapter } from "./adapters/notification-hub/installation";
 import { CreateMassiveNotificationJobUseCase } from "./domain/use-cases/create-massive-notification-job";
+import { HealthCheckUseCase } from "./domain/use-cases/health";
+import { InfoUseCase } from "./domain/use-cases/info";
 import {
   ActivityName as CreateOrUpdateActivityName,
   getActivityHandler as getCreateOrUpdateActivityHandler,
@@ -179,16 +185,28 @@ const main = (config: Config) => {
     queueName: updateInstallationDispatchQueueName,
   });
 
+  const infoUseCase = new InfoUseCase();
   app.http("Info", {
     authLevel: "anonymous",
-    handler: getInfoHandler(
-      apiCosmosdb,
-      pushCosmosDb,
-      blobService,
-      notificationHubClients,
-    ),
+    handler: getInfoHandler(infoUseCase),
     methods: ["GET"],
     route: "api/v1/info",
+  });
+
+  const healthChecks = [
+    () => cosmosHealthcheck(apiCosmosdb),
+    () => cosmosHealthcheck(pushCosmosDb),
+    () => blobServiceHealthcheck(blobService),
+    ...notificationHubClients.map(
+      (client, i) => () => notificationHubHealthcheck(client, i + 1),
+    ),
+  ];
+  const healthCheckUseCase = new HealthCheckUseCase(healthChecks);
+  app.http("Health", {
+    authLevel: "anonymous",
+    handler: getHealthHandler(healthCheckUseCase),
+    methods: ["GET"],
+    route: "api/v1/health",
   });
 
   df.app.orchestration(
@@ -291,9 +309,11 @@ const main = (config: Config) => {
     queueName: updateInstallationDispatchQueueName,
   });
 
-  const massiveJobsRepository = new CosmosMassiveJobsAdapter(
-    pushCosmosDb,
+  const massiveJobsContainer = pushCosmosDb.container(
     config.massiveJobsContainerName,
+  );
+  const massiveJobsRepository = new CosmosMassiveJobsAdapter(
+    massiveJobsContainer,
   );
   const createMassiveNotificationJobUseCase =
     new CreateMassiveNotificationJobUseCase(massiveJobsRepository);