+++ infra/resources/_modules/cosmos/io_com_cosmos.tf
@@ -121,3 +121,16 @@ resource "azurerm_cosmosdb_sql_container" "installation_summaries_lease" {
     max_throughput = 1000
   }
 }
+
+resource "azurerm_cosmosdb_sql_container" "massive_jobs" {
+  name                = "massive-jobs"
+  resource_group_name = var.resource_group
+
+  account_name        = module.io_com_cosmos_account.name
+  database_name       = azurerm_cosmosdb_sql_database.push_notifications.name
+  partition_key_paths = ["/id"]
+
+  autoscale_settings {
+    max_throughput = 2000
+  }
+}